updated navigation to questions page
